% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_endringer_1963}
\alias{api_endringer_1963}
\title{Test data for changes in municipality codes from 1963 onwards}
\format{
An object of class \code{data.frame} with 1172 rows and 7 columns.
}
\usage{
api_endringer_1963
}
\description{
A data.frame containing information about changes in municipality codes
}
\keyword{datasets}
