% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{CorrespondList}
\alias{CorrespondList}
\title{Correspondence list
Print a list of correspondence tables for a given klass with source and target IDs}
\usage{
CorrespondList(klass, date = NULL)
}
\arguments{
\item{klass}{Classification number}

\item{date}{Date for classification (format = "YYYY-mm-dd"). Default is current date}
}
\value{
Data frame with list of corrsepondence tables, source ID and target ID.
}
\description{
Correspondence list
Print a list of correspondence tables for a given klass with source and target IDs
}
\examples{
\dontrun{
CorrespondList("7")
CorrespondList("131", date = "2016-01-01") 
}
}
