% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_equivalents.R
\name{find_equivalent_nodes}
\alias{find_equivalent_nodes}
\title{Find the equivalent sets of a node at various dates}
\usage{
find_equivalent_nodes(node, dates, graph)
}
\arguments{
\item{node}{The node that we're finding the equivalent sets of}

\item{dates}{The dates that we want to find the equivalent sets in}

\item{graph}{The graph that the nodes come from}
}
\value{
A named list of \code{length(dates)} containing the equivalent nodes for
each date. The names of the returned list are the dates provided in \code{date},
coerced to character.
}
\description{
Find the equivalent sets of a node at various dates
}
