% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-build.R
\name{get_klass_changes}
\alias{get_klass_changes}
\title{Get all changes that have occurred in a Klass classification}
\usage{
get_klass_changes(classification)
}
\arguments{
\item{classification}{The ID of the desired classification.}
}
\value{
A \code{data.frame} containing the code changes.
}
\description{
Get all changes that have occurred in a Klass classification
}
