% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{GetFamily}
\alias{GetFamily}
\title{Identify corresponding family from a classification number}
\usage{
GetFamily(klass)
}
\arguments{
\item{klass}{Classification number}
}
\value{
Family number
}
\description{
Identify corresponding family from a classification number
}
\examples{
GetFamily(klass = 7)
}
