% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{GetName}
\alias{GetName}
\title{Get the name of a classification version}
\usage{
GetName(version)
}
\arguments{
\item{version}{Version number}
}
\value{
string or vector of strings with name of version
}
\description{
Get the name of a classification version
}
\examples{
GetName("33")
GetName(GetVersion(family = 1))
}
