% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hent_data.R
\name{GetUrl2}
\alias{GetUrl2}
\title{Get json file from Url - alternative version}
\usage{
GetUrl2(url)
}
\arguments{
\item{url}{String url address}
}
\value{
text in json format
}
\description{
Get json file from Url - alternative version
}
