% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLASS.R
\name{KLASS}
\alias{KLASS}
\title{Match and convert a classification}
\usage{
KLASS(x, klass, date = NULL, correspond = NULL, language = "nb",
  output_level = NULL, output = "name", format = TRUE)
}
\arguments{
\item{x}{Input vector}

\item{klass}{Classification number}

\item{date}{Date for classification (format = "YYYY-mm-dd"). Default is current date}

\item{correspond}{ID number for target in correspondence table. For correspondence between two dates within the same classification, use correspond = TRUE.}

\item{language}{Default "nb" for Norwegian (Bokmål). Also "nn" (Nynorsk) and "en" (English available for some classifications)}

\item{output_level}{Desired output level}

\item{output}{String describing output. May be "name" (default), "code" or "both".}

\item{format}{Run formatting av input vector x (Default = TRUE), important to check if formatting is in one level.}
}
\value{
A vector or data frame is returned with names and/or code of the desired output level.
}
\description{
Match and convert a classification
}
\examples{
load(testdata)
testdata$kommune_names <- KLASS(x = testdata$kommune, klass = 131)
}
