% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klassdata}
\alias{klassdata}
\title{Testdata for KLASS package}
\format{A data frame containing 100 rows and 7 variables:
\describe{
\item{ID}{Identification number}
\item{sex}{1/2 variable for sex}
\item{education}{3- or 5-digit number giving the highest education level using standard education codes. Based on 01.01.2015}
\item{kommune}{4-digit code for Norwegian municipality. Based on 01.01.2015}
\item{kommune2}{Numeric variable for Norwegian municipality with dropped leading zero's for testing. Based on 01.01.2015}
\item{nace5}{5-digit code for industry (NACE). Based on 01.01.2015 standard industry codes}
\item{occupation}{4-digit occupation codes using standard bsed on 01.01.2015}
}}
\usage{
klassdata
}
\description{
A dataset containing variables for testing of Statistics Norways classification API with the KLASS package
}
\keyword{datasets}
