% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-build.R
\name{KlassGraph}
\alias{KlassGraph}
\title{Build a directed graph of code changes based on a Klass classification}
\usage{
KlassGraph(classification, date = NULL)
}
\arguments{
\item{classification}{The ID of the desired classification.}

\item{date}{The date which the edges of the graph should be directed towards.

  Defaults to the current year plus one, which ensures the graph is directed
  to the most recent codes.}
}
\value{
An object containing information about all codes in the
  classification/codelist, and information about the changes between codes.

  The information is stored as a \code{igraph} object with the vertices
  representing codes, and edges representing changes between codes. The
  direction of the edges represent changes towards the date specified in
  \code{date}.
}
\description{
Build a directed graph of code changes based on a Klass classification
}
\examples{

# Build a graph directed towards the most recent codes

klass_131 <- KlassGraph(131)

# Build a graph directed towards valid codes in 2020.

klass_131_2020 <- KlassGraph(131, "2020-01-01")


}
