% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{ListFamily}
\alias{ListFamily}
\title{Classification family list
Print a list of all families and the number of classifications in each}
\usage{
ListFamily(family = NULL, codelists = FALSE)
}
\arguments{
\item{family}{Input family ID number to get a list of classifications in that family}

\item{codelists}{True/False for whether to include codelists. Default = FALSE}
}
\value{
dataset containing a list of families
}
\description{
Classification family list
Print a list of all families and the number of classifications in each
}
\examples{
ListFamily(family = 1)
}
