% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{ListKlass}
\alias{ListKlass}
\title{Classification list
Get a full list of all classifications and codelists}
\usage{
ListKlass(codelists = FALSE)
}
\arguments{
\item{codelists}{True/False for whether to include codelists. Default = FALSE}
}
\value{
A data frame containing a full list of classifications. The data frame includes the classification name, number, family and type.
}
\description{
Classification list
Get a full list of all classifications and codelists
}
\examples{
head(ListKlass(codelists = TRUE))
}
