% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLASS.R
\name{MakeChar}
\alias{MakeChar}
\title{Conversion to character}
\usage{
MakeChar(x)
}
\arguments{
\item{x}{}
}
\value{
x converted to a string or vector of strings.
}
\description{
Conversion to character
}
\examples{
MakeChar(2)
MakeChar(c(2,3,5))
}
