% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{splitChar}
\alias{splitChar}
\title{insert missing dots to the right place in a string}
\usage{
splitChar(x, dot)
}
\arguments{
\item{x}{- character with missing dots}

\item{dot}{- the place of missing dots}
}
\description{
insert missing dots to the right place in a string
}
