% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass.R
\name{UpdateKlass}
\alias{UpdateKlass}
\title{Update multiple Klass codes to a desired date.}
\usage{
UpdateKlass(
  codes,
  dates = NA,
  classification = NULL,
  date = NULL,
  graph = KlassGraph(classification, date),
  output = "code",
  report = FALSE,
  combine = TRUE
)
}
\arguments{
\item{codes}{Codes to be updated.}

\item{dates}{Optional. Can be used to specify what date each of the codes was
valid in. Supply a character vector of length 1 to specify the same valid
date for all codes, or a character vector of the same length as
\code{codes} to specify valid dates for each code. The character vector(s)
should have a format coercible by \code{\link[base]{as.Date}}, e.g.
\code{YYYY-MM-DD}. The function will return an error if a code was not
valid at the specified date.}

\item{classification}{The ID of the desired classification.}

\item{date}{Optional. Can be used to specify the date the codes should be
updated to, e.g. if you have codes that are valid in year \code{T}, but
want to change the codes to the corresponding version in year \code{T-1}.
If unspecified (the default), the function will update codes to the most
recent version.}

\item{graph}{Optional. A graph object generated by \code{\link{KlassGraph}}.
If you're making multiple calls to \code{\link[klassR]{UpdateKlass}}, you
can save some time by generating the graph beforehand and reusing it for
each call to \code{\link[klassR]{UpdateKlass}} with this parameter. If
providing the graph directly, you do not need to provide the
\code{classification} and \code{date} parameters.}

\item{output}{Either a character vector, containing one or more of the items
  in the list below, or \code{TRUE} to include all columns.

  \describe{
   \item{\code{"code"}}{The Klass code.}
   \item{\code{"name"}}{The Klass name.}
   \item{\code{"validFrom"}}{The date that the code is valid from.}
   \item{\code{"validTo"}}{The date that the code is valid to.}
   \item{\code{"split"}}{Logical: Does the code split into two or more codes?}
   \item{\code{"combined"}}{Logical: Does two or more codes become this code?}
   \item{\code{"nextCode"}}{If \code{split == FALSE}, gives the code this code changed into. \code{NA} otherwise.}
  }

  If only one value is specified (e.g. \code{"code"}), the function will
  return either a character vector or a list of character vectors, depending
  on \code{report}, containing the information specified. If more than one
  value is specified (e.g. \code{c("code", "name")}), the function will
  return a list of \code{data.frames} containing the specified columns.}

\item{report}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the output
character vectors/\code{data.frames} will contain information about the
input code, output code, and any "intermediate" codes between the input and
output. If \code{FALSE}, the output will only contain information about the
updated code.}

\item{combine}{\code{TRUE} or \code{FALSE}. If \code{FALSE}, any code that
has been combined with another code will return \code{NA}. If \code{TRUE},
combined codes will not return \code{NA}.}
}
\value{
Either a character vector or list of the same length as \code{codes},
  depending on \code{output} and \code{report}. The function returns
  information about the updated code, and optionally information about
  intermediate codes, as specified by \code{output} and \code{report}.

  Any input codes that have been split will return \code{NA}.
}
\description{
Update multiple Klass codes to a desired date.
}
\examples{

codes <- GetKlass(131, date = "2020-01-01")[["code"]]

updated_codes <- UpdateKlass(codes,
                             dates = "2020-01-01",
                             classification = 131)

}
