% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-navigate.R
\name{UpdateKlassNode}
\alias{UpdateKlassNode}
\title{Given a node and a graph, find the node at the end of a sequence of changes.}
\usage{
UpdateKlassNode(graph, node)
}
\arguments{
\item{graph}{A graph generated by \code{\link{KlassGraph}}.}

\item{node}{A node as returned by \code{\link{KlassNode}} or \code{\link[igraph]{V}}.}
}
\value{
A sequence of vertices, starting with \code{node} and ending with the last
  visited node.
}
\description{
Given a node and a graph, find the node at the end of a sequence of changes.
}
\examples{

# Build a graph directed towards the most recent codes.

klass_131 <- KlassGraph(131)

# Find the most recent node in the graph representing the code "0101" (Halden,
# valid to 2020.)

halden_node <- KlassNode(klass_131, "0101")

# Find the most recent code corresponding to 0101 Halden

halden_node_updated <- UpdateKlassNode(klass_131, halden_node)


}
