% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formattering}
\alias{formattering}
\title{Format vector for industry codes}
\usage{
formattering(x, classification)
}
\arguments{
\item{x}{\itemize{
\item vector of character
}}

\item{classification}{\itemize{
\item classification number
}}
}
\value{
vector of character
}
\description{
Format vector for industry codes
}
\keyword{internal}
