% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formattering}
\alias{formattering}
\title{Convert vector to the right format}
\usage{
formattering(x, input_level, klass, klass_data)
}
\arguments{
\item{x}{- vector of character}

\item{input_level}{- which classification level}

\item{klass}{- classification number}

\item{klass_data}{- the right formatting to the classification levels}
}
\value{
vector of character
}
\description{
Convert vector to the right format
}
\examples{
klass_data <- GetKLASS(klass = "6", date = "2007-01-01")
input_level <- levelCheck(x = klassdata$nace5, klass_data = klass_data)
formattering(x = klassdata$nace5, input_level = input_level, klass = 6, klass_data=klass_data)

klass_data <- GetKLASS(klass = "7", date = "2007-01-01")
input_level <- levelCheck(x = klassdata$occupation, klass_data = klass_data)
formattering(x = klassdata$occupation, input_level = input_level, klass = 7, klass_data=klass_data)

klass_data <- GetKLASS(klass = "131", date = "2007-01-01")
input_level <- levelCheck(x = klassdata$occupation, klass_data = klass_data)
formattering(x = klassdata$kommune2, input_level = input_level, klass = 131, klass_data=klass_data)


}
